redux-toolkit 기본 세팅
