+++ src/App.jsx
@@ -1,0 +1,34 @@
+import { Routes, Route } from "react-router-dom";
+import Main from "./pages/Main";
+import Detail from "./pages/Detail";
+import Create from "./pages/Create";
+import Edit from "./pages/Edit";
+import Signup from "./pages/Signup";
+import Login from "./pages/Login";
+
+function App() {
+  return (
+    // 페이지 이동에 사용되는 Route 태그를 위해선 Routes로 먼저 감싸야 한다.
+    <Routes>
+      {/* path="/"이기 때문에 '<주소>/'인 주소로 접속할 경우 Main 컴포넌트가 화면에 보여지게 된다.  */}
+      {/* props로 items 내려보내주기 */}
+      <Route path="/" element={<Main />} />
+      <Route path="/detail/:id" element={<Detail />} />
+      <Route path="/create" element={<Create />} />
+      <Route path="/edit/:id" element={<Edit />} />
+      <Route path="/signup" element={<Signup />} />
+      <Route path="/login" element={<Login />} />
+      <Route
+        path="*"
+        element={
+          <>
+            <div>없는 페이지입니다.</div>
+            <Link to="/">홈으로 이동</Link>
+          </>
+        }
+      />
+    </Routes>
+  );
+}
+
+export default App;
